requirements.txt
mysql==0.0.3
mysqlclient==2.2.7
